fix: update resource handling and API documentation
- Adjusted descriptions in API documentation to clarify paths for uploads, resources, and containers.
- Updated data models to use consistent naming conventions (e.g., `isDir` to `is_dir`, `modifiedAt` to `modified_at`).
- Changed resource ID types from string to uint64 for better consistency across the application.
- Enhanced flow file upload functionality to support batch processing of resource files.
- Removed deprecated code related to resource entry responses in tests.

+++ backend/pkg/server/services/resources.go
@@ -1544,7 +1544,7 @@ func pathHasTrailingSeparator(p string) bool {
 
 func convertResource(r models.UserResource) models.ResourceEntry {
 	return models.ResourceEntry{
-		ID:        int64(r.ID),
+		ID:        r.ID,
 		UserID:    r.UserID,
 		Name:      r.Name,
 		Path:      r.Path,
@@ -1565,7 +1565,7 @@ func convertResources(recs []models.UserResource) []models.ResourceEntry {
 
 func convertResourceToModel(e models.ResourceEntry) *model.UserResource {
 	return &model.UserResource{
-		ID:        e.ID,
+		ID:        int64(e.ID),
 		UserID:    int64(e.UserID),
 		Name:      e.Name,
 		Path:      e.Path,
